pyproject.toml
[tool.poetry]
name = "lumberjack_auto"
version = "0.1.0"
description = ""
authors = ["homus32 <homus32.reset@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
keyboard = "^0.13.5"
opencv-python = "^4.5.3"
mss = "^6.1.0"
autopy = "^4.0.0"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
